+++ NEWS.md
@@ -22,7 +22,7 @@
 
 * Use Vega-Lite 5.2.0. (#183)
 
-* Update `s3_register()`. (#93)
+* Update `s3_register()`. (#193)
 
 * Remove `schema` from package files, i.e. `system.file("schema", package = "vegawidget")`. 
   Instead, you can use the `vega_schema()` function (with new optional `version`