+++ NEWS.md
@@ -1,4 +1,4 @@
-# vegawidget (development version)
+# vegawidget 0.4.1
 
 * vegawidget now supports the last two Vega-Lite major versions,
   currently versions 5 and 4.